Remove sources in Project folder
